+++ RELEASE_NOTES.md
@@ -2,8 +2,9 @@
 
 ## SC0007
 
-- **TODO** - Fill me in...
+- Update database schema changes: 
 
+        mysql -u root hotornot-dev < VolleyBackend/bin/data_schemas/hotornot-dev-schema-sc0007-00.sql
 
 
 ## SC0006